+++ modules/resource-manager/src/test/java/com/ibm/cloud/platform_services/resource_manager/v2/model/CreateResourceGroupOptionsTest.java
@@ -1,0 +1,44 @@
+/*
+ * (C) Copyright IBM Corp. 2020.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ */
+
+package com.ibm.cloud.platform_services.resource_manager.v2.model;
+
+import com.ibm.cloud.platform_services.resource_manager.v2.model.CreateResourceGroupOptions;
+import com.ibm.cloud.platform_services.resource_manager.v2.utils.TestUtilities;
+
+import com.ibm.cloud.sdk.core.service.model.FileWithMetadata;
+import java.io.InputStream;
+
+import java.util.HashMap;
+import java.util.List;
+
+import org.testng.annotations.Test;
+import static org.testng.Assert.*;
+
+/**
+ * Unit test class for the CreateResourceGroupOptions model.
+ */
+public class CreateResourceGroupOptionsTest {
+  final HashMap<String, InputStream> mockStreamMap = TestUtilities.createMockStreamMap();
+  final List<FileWithMetadata> mockListFileWithMetadata = TestUtilities.creatMockListFileWithMetadata();
+
+  @Test
+  public void testCreateResourceGroupOptions() throws Throwable {
+    CreateResourceGroupOptions createResourceGroupOptionsModel = new CreateResourceGroupOptions.Builder()
+      .name("test1")
+      .accountId("25eba2a9-beef-450b-82cf-f5ad5e36c6dd")
+      .build();
+    assertEquals(createResourceGroupOptionsModel.name(), "test1");
+    assertEquals(createResourceGroupOptionsModel.accountId(), "25eba2a9-beef-450b-82cf-f5ad5e36c6dd");
+  }
+}