+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## [0.4.1](https://github.com/IBM/platform-services-java-sdk/compare/0.4.0...0.4.1) (2020-05-15)
+
+
+### Bug Fixes
+
+* **build:** disable integration tests in tagged-release builds ([#22](https://github.com/IBM/platform-services-java-sdk/issues/22)) ([30f0afe](https://github.com/IBM/platform-services-java-sdk/commit/30f0afef3368fbf03531867c564f402656533eb5))
+
+# [0.4.0](https://github.com/IBM/platform-services-java-sdk/compare/0.3.0...0.4.0) (2020-05-12)
+
+
+### Features
+
+* **Resource Manager:** add service to project ([#15](https://github.com/IBM/platform-services-java-sdk/issues/15)) ([df43198](https://github.com/IBM/platform-services-java-sdk/commit/df43198af1b5a57ac1730f4ac7a24da2aed6d717))
+
 # [0.3.0](https://github.com/IBM/platform-services-java-sdk/compare/0.2.2...0.3.0) (2020-05-07)